* Renamed Encoder and Decoder classes
